+++ CHANGES.md
@@ -1,6 +1,13 @@
 Zing Changelog
 ==============
 
+v.next
+------
+
+* Editor: fixed a bug where the Plurr editor would go away and miss its state
+  (#309).
+
+
 v0.8.0 (2018-01-29)
 -------------------